requirements.txt
fastapi
uvicorn
transformers
torch
faiss-cpu
requests
